requirements.txt
python-binance
pandas
pandas-ta
python-dotenv